fixbug connect computer with account

+++ src/main/java/vn/eledevo/vksbe/dto/response/Result.java
@@ -12,5 +12,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Result<T> {
     List<T> content;
-    Long total;
+    Integer total;
 }
